1) insert measure works again 2) backspace works properly 3) note stems work properly 4) note lengths are in ints, not longs 5) rests are fixed up better
regression introduced - cursor goes to the wrong place after note inserted

+++ gui/src/idiocy/ui/data/TimeSig.scala
@@ -10,5 +10,5 @@ object TimeSig{
 }
 
 case class TimeSig(top: Int, bottom: Int) {
-  def measureLengthPips: Long = MusicEvent.PipsToABeat * 4 * top / bottom
+  def measureLengthPips: Int = MusicEvent.PipsToABeat * 4 * top / bottom
 }
